+++ django_tables2/columns/templatecolumn.py
@@ -3,6 +3,8 @@ from __future__ import absolute_import, unicode_literals
 
 from django.template import Context, Template
 from django.template.loader import get_template
+from django.utils import six
+from django.utils.html import strip_tags
 
 from .base import Column, library
 
@@ -20,9 +22,10 @@ class TemplateColumn(Column):
     A `~django.template.Template` object is created from the
     *template_code* or *template_name* and rendered with a context containing:
 
-    - *record*  -- data record for the current row
-    - *value*   -- value from `record` that corresponds to the current column
-    - *default* -- appropriate default value to use as fallback
+    - *record*      -- data record for the current row
+    - *value*       -- value from `record` that corresponds to the current column
+    - *default*     -- appropriate default value to use as fallback
+    - *row_counter* -- The number of the row this cell is being rendered in.
 
     Example:
 
@@ -53,7 +56,8 @@ class TemplateColumn(Column):
             'default': bound_column.default,
             'column': bound_column,
             'record': record,
-            'value': value
+            'value': value,
+            'row_counter': kwargs['bound_row'].row_counter
         })
 
         try:
@@ -63,3 +67,14 @@ class TemplateColumn(Column):
                 return get_template(self.template_name).render(context.flatten())
         finally:
             context.pop()
+
+    def value(self, **kwargs):
+        '''
+        The value returned from a call to `value()` on a `TemplateColumn` is
+        the rendered tamplate with `django.utils.html.strip_tags` applied.
+        '''
+        html = super(TemplateColumn, self).value(**kwargs)
+        if isinstance(html, six.string_types):
+            return strip_tags(html)
+        else:
+            return html