+++ django_tables2/templatetags/django_tables2.py
@@ -50,10 +50,11 @@ def token_kwargs(bits, parser):
 
 
 class QuerystringNode(Node):
-    def __init__(self, updates, removals):
+    def __init__(self, updates, removals, asvar=None):
         super(QuerystringNode, self).__init__()
         self.updates = updates
         self.removals = removals
+        self.asvar = asvar
 
     def render(self, context):
         if 'request' not in context:
@@ -67,10 +68,17 @@ class QuerystringNode(Node):
                 params[key] = value
         for removal in self.removals:
             params.pop(removal.resolve(context), None)
-        return escape('?' + urlencode(params, doseq=True))
+
+        value = escape('?' + urlencode(params, doseq=True))
+
+        if self.asvar:
+            context[str(self.asvar)] = value
+            return ''
+        else:
+            return value
 
 
-# {% querystring "name"="abc" "age"=15 %}
+# {% querystring "name"="abc" "age"=15 as=qs %}
 @register.tag
 def querystring(parser, token):
     '''
@@ -80,21 +88,34 @@ def querystring(parser, token):
 
     Example (imagine URL is ``/abc/?gender=male&name=Brad``)::
 
+        # {% querystring "name"="abc" "age"=15 %}
+        ?name=abc&gender=male&age=15
         {% querystring "name"="Ayers" "age"=20 %}
         ?name=Ayers&gender=male&age=20
         {% querystring "name"="Ayers" without "gender" %}
         ?name=Ayers
-
     '''
     bits = token.split_contents()
     tag = bits.pop(0)
     updates = token_kwargs(bits, parser)
+
+    asvar_key = None
+    for key in updates:
+        if str(key) == 'as':
+            asvar_key = key
+
+    if asvar_key is not None:
+        asvar = updates[asvar_key]
+        del updates[asvar_key]
+    else:
+        asvar = None
+
     # ``bits`` should now be empty of a=b pairs, it should either be empty, or
     # have ``without`` arguments.
     if bits and bits.pop(0) != 'without':
         raise TemplateSyntaxError("Malformed arguments to '%s'" % tag)
     removals = [parser.compile_filter(bit) for bit in bits]
-    return QuerystringNode(updates, removals)
+    return QuerystringNode(updates, removals, asvar=asvar)
 
 
 class RenderTableNode(Node):